Shorten opencl kernel params, add source file for openclKernels
